requirements.txt
streamlit>=1.30
langchain>=0.2
langchain-google-genai>=1.0
google-generativeai>=0.5
folium>=0.15
streamlit-folium>=0.15
twilio>=8.0
requests>=2.31
python-dotenv>=1.0
geopy>=2.4
